+++ src/powerups/timewarp.cr
@@ -9,7 +9,7 @@ class PowerupTimeWarp < Powerup
   DURATION = 600
   KEY_DURATION = "timewarp_duration"
 
-  def new_multiplier(public_key) : Float64
+  def new_multiplier(public_key) : BigFloat
     get_synergy_boosted_multiplier(public_key, UNIT_MULTIPLIER)
   end
 
@@ -35,7 +35,7 @@ class PowerupTimeWarp < Powerup
   end
 
   def get_description(public_key)
-    
+
     amount = ((get_unit_boost(public_key)) * new_multiplier(public_key)).round(2)
     "Multiplies unit production by #{amount}x for the next 10 minutes. Price increases exponentially."
   end
@@ -86,7 +86,7 @@ class PowerupTimeWarp < Powerup
     durations.each do |t|
       boost_units *= t[1].to_f
     end
-    
+
     boost_units
   end
